+++ app/src/main/kotlin/org/cru/godtools/ui/settings/language/downloadable/LanguageDownloadStatusIndicator.kt
@@ -1,4 +1,4 @@
-package org.cru.godtools.ui.languages.downloadable
+package org.cru.godtools.ui.settings.language.downloadable
 
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.foundation.border
