+++ app/src/test/snapshots/images/org.cru.godtools.ui.settings.language.downloadable_DownloadableLanguagesLayoutPaparazziTest_DownloadableLanguagesLayout()_-_Searching[Nexus_5,NIGHT,NO_ACCESSIBILITY].png
