+++ app/src/test/snapshots/images/org.cru.godtools.ui.settings.language.downloadable_DownloadableLanguagesLayoutPaparazziTest_DownloadableLanguagesLayout()_-_Searching[Nexus_5,NOTNIGHT,NO_ACCESSIBILITY].png
