+++ app/src/test/snapshots/images/org.cru.godtools.ui.settings.language.downloadable_LanguageDownloadStatusIndicatorPaparazziTest_LanguageDownloadStatusIndicator()[NOTNIGHT].png
